Add Djscovery logo and align login elements
